Fix: pull을 issue로 네이밍 변경 적용
